+++ CHANGELOG.md
@@ -4,10 +4,10 @@
 
 ### Changes
 
-- Added GuestHole support, that can be used by the clients to overcome potential blocked api. Mainly
-this is supported now from VPN client.
+- Added Server Connection error support, that can be used by the clients to overcome potential blocked api.
+Mainly and initially this is supported now from VPN client for their guest hole feature.
 - Clients should provide a new dependency for `GuestHoleFallbackListener` or use the default one that
-Core provides `DefaultGustHoleFallbackListener` which does not do anything.
+Core provides `ServerConnectionListener` which does not do anything.
 
 ## Auth [1.18.4], Network [1.15.8], Util Kotlin [1.15.3]
 ## Auth [1.18.5], Network [1.15.9], Util Kotlin [1.15.4]